requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
gitdb2==2.0.5
GitPython==2.1.11
idna==2.8
lxml==4.3.2
MechanicalSoup==0.11.0
requests==2.21.0
ruamel.yaml==0.15.89
six==1.12.0
smmap2==2.0.5
soupsieve==1.8
urllib3==1.24.1
wincertstore==0.2
